requirements.txt
telethon
python-dotenv
asyncio
pandas
sqlalchemy
dbt-postgres
ultralytics
psycopg2-binary 
pydantic
fastapi 
uvicorn